requirements.txt
Flask==1.1.1
mysqlclient==1.4.2.post1
python-dotenv==0.10.1
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
pika==1.1.0
dialogflow==0.7.2
fasttext==0.9.1
celery==4.4.0
redis==3.3.11
flask-admin==1.5.4
flask-security==3.0.0
bcrypt==3.1.7
python-telegram-bot==12.2.0
pyTelegramBotAPI==3.6.6
pyenchant==2.0.0
guess-language-spirit==0.5.3
yandex.translate==0.3.5